Stop installing openstack-tempest with designate-tempest-plugin
openstack-tempest is packaged in RDO from a fork in
redhat-openstack/tempest and this fork is not currently well
maintained.
As of now:
- This branch is 192 commits ahead, 324 commits behind openstack:master.

For this reason, some consumers that deploy and test from RDO
packages prefer to install and use Tempest directly from source.

Installing openstack-tempest alongside a tempest installed from
source leads to a lot of weird errors when running the tests.

If a user wants to install tempest, he can explicitely install
openstack-tempest or install it from source. Additionnally,
openstack-tempest is not currently a requirement when installing
anything of the "python-<project>-tests" packages.

Change-Id: I00e7f1795b4317987550fab44463b872029df5be