Update patch set 3
Patch Set 3:

I just want to highlight some points on tempest:
Currently upstream tempest is brachless and is suppose to work with all the supported version of tempest.
The All openstack services tempest plugins are branched since it is resides with in the openstack service repos and in order to work a tempest plugin fully with tempest we might need all the tempest plugins.

Patch-set: 3