Update patch set 3
Patch Set 3: Code-Review+1

In an ideal world, this package should depend on openstack-tempest (after all, it's needed for the designate tempest tests to be useful at all). However, given our current packaging, it makes sense to do it this way even if just for consistency, so +1 for me (don't have +2 in this project :?).

BTW, https://review.rdoproject.org/r/1780 is adding the dependency the other way around, so let's keep discussing until we find the right fix.

Patch-set: 3
Label: Code-Review=+1