Update patch set 3
Patch Set 3:

The release candidates are almost around the corner and I would really like to be able to say we have designate test coverage. It was (and still is) utterly broken.

Some of the work that has been done so far:
- https://review.openstack.org/#/q/topic:designate-with-rdo
- https://review.rdoproject.org/r/#/q/topic:designate-with-rdo

Keeping a package broken on a technicality that can be addressed later if we change our mind is frustrating
I'm going to re-iterate that none of the other tempest plugins currently install openstack-tempest, some examples: 

- python-nova-tests: https://github.com/rdo-packages/nova-distgit/blob/rpm-master/openstack-nova.spec#L424
- python-glance-tests: https://github.com/rdo-packages/glance-distgit/blob/rpm-master/openstack-glance.spec#L174
- python-horizon-tests-tempest: https://github.com/rdo-packages/tempest-horizon-distgit/blob/rpm-master/python-horizon-tests-tempest.spec
- python-aodh-tests: https://github.com/rdo-packages/aodh-distgit/blob/rpm-master/openstack-aodh.spec#L207

Patch-set: 3
Reviewer: Gerrit User 16 <16@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0